Docs: Warn about twilight task deadlock
Potential deadlock (identified by a user) has now been warned about. The way the example is structured prevents this from occurring, but it's worth making this more explicit due to the more free-form nature of twilight.

The design of serenity's event handling and framework should prevent this issue from cropping up when using it as a gateway backend.