Working MP3 playback.
MP3s now work great under the convert -> resample -> mix pipeline, even across unclean frame boundaries. Main issue seemed to be the number of internal subchunks in the resampler, which is still totally opaque... Oh well.

Actual instantiation of Lazy->Wrapped and Wrapped->Parsed elements in the driver is not yet covered.

Other formats are broken due to handling of the default track id, which is currently a) messy, and b) incorrect.