Fix Symphonia to v0.5
This version includes all the API changes we really need to function, and adds MKV support to a satisfactory level. This should free this PR up for an actual release down the line.

Additionally, this changes the symphonia dependency to remove default features so that choice over the codec suite (barring opus) is entirely user-driven.