Manually check for backseeking with one-way Inputs
It seems that some formats (MP4, Ogg) don't have logic to check that you're about to make a Very Bad Decision. We need to cache the inner bytestream's `is_seekable` and check this against a backwards seek ourselves, since pulling it out of a constructed source is nigh-on impossible.

MKV/WebM seek bug is still present: so far as I can tell it's because it uses a very different time base in the format than the decoder instances. I need to confirm whether other codecs do any local time_base corrections.