Rework `add_raw` into `add_`/`enqueue_with_preload`.
This does a little bit of internal reworking to make this feasible, but this prevents us from having one function on `TrackHandles` while the rest are on `Track`s (so more user-friendly). Similarly, this prevents us from having to impose conditions on the target `TrackHandle` -- "you must pause your track first" is a bit of a footgun.

This also applies event listeners *before* tracks are handed over to the driver, which is more correct and means we don't need to wait a tick or similar.